pyproject.toml
[project]
name = "learnrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "torchaudio>=2.6.0",
    "pillow>=11.1.0",
    "huggingface-hub>=0.28.1",
    "ipython>=8.32.0",
    "matplotlib>=3.10.0",
    "jupyter>=1.1.1",
    "sqlalchemy>=2.0.38",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.10",
    "pyexiftool>=0.5.6",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
